Refactored code: renamed classes, updated klondikeField to TVector<deck>, added const to methods, fixed deckRenderer error, resolved compiler warnings, replaced enum with enum class, added WGP namespace, changed naming style to camelCase, created CHANGELOG.md.

+++ src/core/fundamental classes/TVector/TVector.h
@@ -199,7 +199,7 @@ template<class T> T* TVector<T>::data() const noexcept {
 }
 
 template<class T> int TVector<T>::size() const noexcept {
-    return _size - _deleted;
+    return static_cast<int>(_size - _deleted);
 }
 
 template<class T> int TVector<T>::capacity() const noexcept {
